Error solver refactor
